Code Cleanup
Why Was This Change Necessary?

* The code was badly-indented, and harder to read.
* There were incorrect use of globals
* The constructor creation was not optimal.

How Does It Address the Problem?

By…

* Fixing the indentation;
* Properly declaring variables with `var`;
* Renaming certain variables to increase readability;
* Replacing `escape` with `decodeURIComponent`
(which is the suggested way of escaping cookie values);
* Returning early by using guard clauses instead of
keeping code nested in a chain of `if/else`s
* Some other minor impromenents.

Are There Any Side Effects?

I haven’t tested the code, however it should be good to go
for most use cases; and if there are any problems, it
should be relatively easy to apply a patch.

Any suggestions?

I am sure this code is great for a learning experience;
and also it can fit simple use cases.
However, it would be nice to mention solutions like
<https://github.com/js-cookie/js-cookie> which have been battle-tested
which has already been battle-tested to the ground, and has been used
by masses, edge-cases polished, and runs in almost any platform.

It might even make sense to fork `js-cookie` and modify/simplify it
as an alternative (if “simplification" is the goal here, as the name suggests).
or maybe contribute to the project instead of reinventing the wheel alone.

I would also suggest:

* Adding unit tests to the code (since it is damn easy to do);
* And Writing a proper README that at least show some example use cases.

At any rate, welcome to the open-source world buddy!

Enjoy your ride!